import toilet, check scale, rescale toilet, minor tweak to test scene, test material folder
I noticed the toilet model is far too small relative to the character controller/player height. I've eyeballed it with a scale shape to reference in blender. Perhaps we should consider a central scale reference, otherwise we will have to eyeball all models to the character controller height as it currently is. I added some colours to the test scene so that I can see properly in unshaded view.