Fixed weird issue related to paints being dark at some distance of the camera.
The issue was triggered by transparent material rendering order, which are used both in the corridor walls textures and in the paint and the wall textures and at some distance the wall was rendered over the paint instead of the reverse. Fixed by making the paint material rendering higher priority.