treasure chest anim fix
I've tweaked the treasure chest files, and exported a fixed version for game engine. There is a lid-opening animation that works when interacted with. Some kind of issue with blender exports, I'll have to look into blender gltf exports for Godot in future, it doesn't seem like it's all worked out properly...